add GridHelper and complated the Related Image(but still have problem)
