+++ app/src/main/java/com/lukitateam/lukita/ui/components/GridHelper.kt
@@ -1,0 +1,63 @@
+@file:OptIn(ExperimentalFoundationApi::class)
+
+package com.lukitateam.lukita.ui.components
+
+import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.lazy.staggeredgrid.LazyVerticalStaggeredGrid
+import androidx.compose.foundation.lazy.staggeredgrid.StaggeredGridCells
+import androidx.compose.foundation.lazy.staggeredgrid.items
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+import kotlin.random.Random
+
+@Composable
+fun GridHelper() {
+    LazyVerticalStaggeredGrid(
+        columns = StaggeredGridCells.Adaptive(128.dp),
+        modifier = Modifier.fillMaxSize(),
+        contentPadding = PaddingValues(16.dp),
+        horizontalArrangement = Arrangement.spacedBy(16.dp),
+        verticalArrangement = Arrangement.spacedBy(16.dp)
+    ) {
+        items(items) { item ->
+            RandomColorBox(item = item)
+        }
+    }
+}
+
+val items = (1..100).map {
+    ListItem(
+        height = Random.nextInt(100, 300).dp,
+        color = Color(
+            Random.nextLong(0xFFFFFFFF)
+        ).copy(alpha = 1f)
+    )
+}
+
+data class ListItem(
+    val height: Dp,
+    val color: Color,
+)
+
+@Composable
+fun RandomColorBox(item: ListItem) {
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(item.height)
+            .clip(RoundedCornerShape(10.dp))
+            .background(item.color)
+    )
+}
